+++ src/General/Exceptions/InvalidUrlException.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Pageon\SlackWebhookMonolog\Slack\Exceptions;
+namespace Pageon\SlackWebhookMonolog\General\Exceptions;
 
 use InvalidArgumentException;
 
 /**
  * @author Jelmer Prins <jelmer@pageon.be>
  *
- * @since 0.1.0
+ * @since 0.3.2
  */
 class InvalidUrlException extends InvalidArgumentException
 {