+++ src/General/SerializeToString.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pageon\SlackWebhookMonolog\Slack;
+namespace Pageon\SlackWebhookMonolog\General;
 
 use JsonSerializable;
 
@@ -8,6 +8,8 @@ use JsonSerializable;
  * This will return the string representation when the class is encoded to json.
  *
  * @author Jelmer Prins <jelmer@pageon.be>
+ *
+ * @since 0.3.2
  */
 abstract class SerializeToString implements JsonSerializable
 {